==> 10_participant_summary.Rmd <==
---
title: "DataFest at [HOST] [YEAR] - Teams and participants"
output: 
  html_document: 
    theme: cerulean
---

```{r load-packages, echo=FALSE, message=FALSE}
library(readr)
library(DT)
library(knitr)
library(dplyr)
library(ggplot2)
library(stringr)
```

```{r hide-code, echo=FALSE}
opts_chunk$set(echo=FALSE)
```

```{r load-data, message=FALSE}
participants <- read_csv("data/participants.csv")
```

## Overall

```{r}
n_participant <- nrow(participants)
n_team <- length(unique(participants$team_name))
today <- Sys.Date()
```

As of `r today`:

- Number of participants: `r n_participant`

- Number of teams: `r n_team`

## Schools

```{r}
participants %>%
  count(school) %>%
  arrange(desc(n)) %>%
    datatable(options = list(
    lengthChange = FALSE
    ))
```


# Schools and teams

```{r}
participants %>%
  distinct(team_name, .keep_all = TRUE) %>%
  count(school) %>%
  arrange(desc(n)) %>%
  datatable(options = list(
    lengthChange = FALSE
    ))
```

# Year

```{r}
participants <- participants %>%
  mutate(class_year = factor(class_year, levels = c("First-year",
                                                    "Sophomore",
                                                    "Junior",
                                                    "Senior")))
```


### All years

```{r}
participants %>%
  count(class_year) %>%
  datatable(options = list(
    lengthChange = FALSE
    ))
```

### Schools and year

```{r}
ggplot(data = participants, aes(x = school, fill = class_year)) +
  geom_bar(position = "fill")
```

# Majors

```{r}
participants <- participants %>%
  mutate(
    major = tolower(major), 
    is_stat = str_detect(major, "stat"), 
    is_math = str_detect(major, "math"), 
    is_cs = str_detect(major, "computer"), 
    is_eng = str_detect(major, "eng"), 
    is_ns = str_detect(major, "natural"), 
    is_ss = str_detect(major, "social"), 
    is_other = !(is_stat | is_math | is_cs | is_eng | is_ns | is_ss)
  )
```

The values below add up to more than 100% since some students have double majors.

```{r fun}
calc_major_perc <- function(data, is_major){
  major_perc <- participants %>%
    group_by_(is_major) %>%
    summarise(n = n()) %>%
    mutate(perc = n / sum(n)) %>%
    slice(2) %>%
    select(perc) %>%
    as.numeric() %>%
    round(2) * 100
  return(major_perc)
}
```

```{r}
other_majors <- unique(participants$major[participants$is_other == TRUE])
```

- Statistics: `r calc_major_perc(participants, "is_stat")`%
- Computer Science: `r calc_major_perc(participants, "is_cs")`%
- Mathematics: `r calc_major_perc(participants, "is_math")`%
- Engineering: `r calc_major_perc(participants, "is_eng")`%
- Natural Sciences: `r calc_major_perc(participants, "is_ns")`%
- Social Sciences: `r calc_major_perc(participants, "is_ss")`%
- Other: `r calc_major_perc(participants, "is_other")`% (These include `r other_majors`.)

# Team sizes

```{r}
avg_team_size <- participants %>%
  group_by(team_name) %>%
  summarise(team_size = n()) %>%
  summarise(avg_team_size = mean(team_size)) %>%
  as.numeric %>%
  round(2)
```

Average team size is `r avg_team_size`.

```{r}
participants %>%
  count(team_name) %>%
  arrange(n, team_name) %>%
  datatable()
```

# Participants

```{r}
participants %>%
  select(team_name, school, first_name, last_name) %>%
  arrange(team_name, last_name, first_name) %>%
  datatable(options = list(
    pageLength = 30,
    lengthMenu = c(10, 30, 50)
    ))
```
